change status code for error messages

+++ src/main/g8/$seedName$/src/main/scala/$package__packaged$/domain/template/$className;format="Camel"$Service.scala
@@ -25,7 +25,7 @@ class $className;format="Camel"$Service(repo: $className;format="Camel"$Reposito
       .map(_.map(_.transform).asRight[Error])
       .recover { case e: Throwable =>
         logger.error("Error occurred during loading tweets", e)
-        Error(StatusCodes.InternalServerError, "Twitter api is not available").asLeft
+        Error(StatusCodes.InternalServerError.intValue, "Twitter api is not available").asLeft
       }
   }
